+++ bestmobabot/bot.py
@@ -328,7 +328,11 @@ class Bot(contextlib.AbstractContextManager, BotHelperMixin):
         Тренирует предсказательную модель для арены.
         """
         logger.info('Running trainer…')
-        Trainer(self.db, n_splits=constants.MODEL_N_SPLITS).train(params=self.settings.bot.arena.hyper_params)
+        Trainer(
+            self.db,
+            n_splits=constants.MODEL_N_SPLITS,
+            n_last_battles=self.settings.bot.arena.last_battles,
+        ).train()
 
     def attack_arena(self):
         """