+++ bestmobabot/trainer.py
@@ -11,8 +11,20 @@ from bestmobabot.model import Trainer
 
 @click.command()
 @click.option('verbosity', '-v', '--verbose', count=True, help='Increase verbosity.')
-@click.option('--n-splits', type=int, default=constants.MODEL_N_SPLITS, help='K-fold splits.')
-def main(verbosity: int, n_splits: int):
+@click.option(
+    '--n-splits',
+    type=int,
+    default=constants.MODEL_N_SPLITS, help='K-fold splits.',
+    show_default=True,
+)
+@click.option(
+    '--n-last-battles',
+    type=int,
+    default=constants.MODEL_N_LAST_BATTLES,
+    help='Use N last battles for training.',
+    show_default=True,
+)
+def main(verbosity: int, n_splits: int, n_last_battles: int):
     """
     Train and generate arena prediction model.
     """
@@ -21,7 +33,7 @@ def main(verbosity: int, n_splits: int):
         warnings.simplefilter('ignore')
 
     with Database(constants.DATABASE_NAME) as db:
-        Trainer(db, n_splits=n_splits).train()
+        Trainer(db, n_splits=n_splits, n_last_battles=n_last_battles).train()
 
 
 if __name__ == '__main__':